refactor(common): 패키지 구조 재구성
