--- academic-service/src/main/java/com/green/academic/application/controller/CalandarController.java
@@ -1,9 +1,0 @@
-package com.green.academic.application.controller;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/grades")
-public class CalandarController {
-}
