--- academic-service/src/main/java/com/green/academic/application/model/calandar/ScheduleListRes.java
@@ -1,4 +1,0 @@
-package com.green.academic.application.model.calandar;
-
-public class ScheduleListRes {
-}
