+++ academic-service/src/main/java/com/green/academic/application/schedule/ScheduleController.java
@@ -1,0 +1,9 @@
+package com.green.academic.application.schedule;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/admin/lectures")
+public class ScheduleController {
+}
