+++ academic-service/src/main/java/com/green/academic/application/schedule/ScheduleService.java
@@ -1,0 +1,7 @@
+package com.green.academic.application.schedule;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class ScheduleService {
+}
