+++ academic-service/src/main/java/com/green/academic/application/schedule/model/ScheduleListRes.java
@@ -1,0 +1,4 @@
+package com.green.academic.application.schedule.model;
+
+public class ScheduleListRes {
+}
