--- academic-service/src/main/java/com/green/academic/application/service/CalandarService.java
@@ -1,7 +1,0 @@
-package com.green.academic.application.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class CalandarService {
-}
