--- academic-service/src/main/java/com/green/academic/entity/AcademicCalandars.java
@@ -1,4 +1,0 @@
-package com.green.academic.entity;
-
-public class AcademicCalandars {
-}
