+++ academic-service/src/main/java/com/green/academic/entity/Schedule.java
@@ -1,0 +1,4 @@
+package com.green.academic.entity;
+
+public class Schedule {
+}
