+++ auth-service/src/main/java/com/green/auth/application/auth/AuthMemberRepository.java
@@ -1,0 +1,8 @@
+package com.green.auth.application.auth;
+
+import com.green.auth.entity.AuthMember;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AuthMemberRepository extends JpaRepository<AuthMember, Integer> {
+    boolean existsByMemberCodeAndEmail(Integer memberCode, String email);
+}
