--- auth-service/src/main/java/com/green/auth/repository/AuthMemberRepository.java
@@ -1,8 +1,0 @@
-package com.green.auth.repository;
-
-import com.green.auth.entity.AuthMember;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AuthMemberRepository extends JpaRepository<AuthMember, Integer> {
-    boolean existsByMemberCodeAndEmail(Integer memberCode, String email);
-}
