--- core-service/src/main/java/com/green/core/application/admin/AdminLectureController.java
@@ -1,9 +1,0 @@
-package com.green.core.application.admin;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/admin/lectures")
-public class AdminLectureController {
-}
