--- core-service/src/main/java/com/green/core/application/controller/LectureController.java
@@ -1,9 +1,0 @@
-package com.green.core.application.controller;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/lectures")
-public class LectureController {
-}
