+++ core-service/src/main/java/com/green/core/application/grade/GradeController.java
@@ -1,0 +1,9 @@
+package com.green.core.application.grade;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/calendars")
+public class GradeController {
+}
