+++ core-service/src/main/java/com/green/core/application/grade/GradeService.java
@@ -1,0 +1,7 @@
+package com.green.core.application.grade;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class GradeService {
+}
