+++ core-service/src/main/java/com/green/core/application/lecture/AdminLectureController.java
@@ -1,0 +1,9 @@
+package com.green.core.application.lecture;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/admin/lectures")
+public class AdminLectureController {
+}
