+++ core-service/src/main/java/com/green/core/application/lecture/LectureController.java
@@ -1,0 +1,13 @@
+package com.green.core.application.lecture;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Slf4j
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/lectures")
+public class LectureController {
+}
