+++ core-service/src/main/java/com/green/core/application/lecture/LectureService.java
@@ -1,0 +1,7 @@
+package com.green.core.application.lecture;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class LectureService {
+}
