+++ core-service/src/main/java/com/green/core/application/lecture/model/LectureCreateReq.java
@@ -1,0 +1,4 @@
+package com.green.core.application.lecture.model;
+
+public class LectureCreateReq {
+}
