+++ core-service/src/main/java/com/green/core/application/major/AdminMajorController.java
@@ -1,0 +1,28 @@
+package com.green.core.application.major;
+
+import com.green.common.model.ResultResponse;
+import com.green.core.application.major.model.MajorCreateReq;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Slf4j
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/admin/majors")
+public class AdminMajorController {
+    private final MajorService majorService;
+
+    @PostMapping
+    public ResultResponse<?> test(@RequestBody MajorCreateReq req ) {
+        log.info("req: {}", req);
+        majorService.test( req );
+        return ResultResponse.builder()
+                .message( "학과 생성 테스트 성공" )
+                .data( 1 )
+                .build();
+    }
+}
