+++ core-service/src/main/java/com/green/core/application/major/MajorController.java
@@ -1,0 +1,15 @@
+package com.green.core.application.major;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Slf4j
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/majors")
+public class MajorController {
+    private final MajorService majorService;
+
+}
