+++ core-service/src/main/java/com/green/core/application/major/MajorRepository.java
@@ -1,0 +1,7 @@
+package com.green.core.application.major;
+
+import com.green.core.entity.Major;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MajorRepository extends JpaRepository<Major, Integer> {
+}
