+++ core-service/src/main/java/com/green/core/application/major/model/MajorCreateReq.java
@@ -1,0 +1,8 @@
+package com.green.core.application.major.model;
+
+import lombok.Data;
+
+@Data
+public class MajorCreateReq {
+    private String name;
+}
