--- core-service/src/main/java/com/green/core/application/model/lecture/LectureCreateReq.java
@@ -1,4 +1,0 @@
-package com.green.core.application.model.lecture;
-
-public class LectureCreateReq {
-}
