--- core-service/src/main/java/com/green/core/application/model/major/MajorCreateReq.java
@@ -1,8 +1,0 @@
-package com.green.core.application.model.major;
-
-import lombok.Data;
-
-@Data
-public class MajorCreateReq {
-    private String name;
-}
