--- core-service/src/main/java/com/green/core/application/service/GradeService.java
@@ -1,7 +1,0 @@
-package com.green.core.application.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class GradeService {
-}
