--- core-service/src/main/java/com/green/core/repository/MajorRepository.java
@@ -1,7 +1,0 @@
-package com.green.core.repository;
-
-import com.green.core.entity.Major;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MajorRepository extends JpaRepository<Major, Integer> {
-}
