+++ core-service/src/main/resources/application.yaml
@@ -1,11 +1,8 @@
 constants:
   api:
     prefix: /api/core
-
-
 server:
   port: ${CORE_PORT}
-
 spring:
   config:
     import:
