+++ member-service/src/main/java/com/green/member/application/MemberController.java
@@ -1,0 +1,11 @@
+package com.green.member.application;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.RestController;
+
+@Slf4j
+@RestController
+@RequiredArgsConstructor
+public class MemberController {
+}
