+++ member-service/src/main/java/com/green/member/application/StudentRepository.java
@@ -1,0 +1,7 @@
+package com.green.member.application;
+
+import com.green.member.entity.Student;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface StudentRepository extends JpaRepository<Student, Long> {
+}
