--- member-service/src/main/java/com/green/member/repository/StudentRepository.java
@@ -1,7 +1,0 @@
-package com.green.member.repository;
-
-import com.green.member.entity.Student;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StudentRepository extends JpaRepository<Student, Long> {
-}
